Revert "Прошли тесты постман для всей сборки"
This reverts commit fa18b072adf86967778fe9d8a2c281c2be020607.

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.filmorate.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.dal.EventRepository;
 import ru.yandex.practicum.filmorate.dal.UserRepository;
 import ru.yandex.practicum.filmorate.exception.InvalidFormatException;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
@@ -12,8 +13,8 @@ import ru.yandex.practicum.filmorate.model.Operation;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.Collection;
-import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 
 @Slf4j
@@ -22,13 +23,11 @@ import java.util.Optional;
 public class UserService {
 
     private final UserRepository userRepository;
-    private final EventService eventService;
+    private final EventRepository eventRepository;
 
 
     public Collection<Event> getAllEventsByUserId(int userId) {
-        throwIfNotExists(userId);
-
-        return eventService.findAll(userId);
+        return eventRepository.findAll(userId);
     }
 
     public void removeUser(int id) {
@@ -48,7 +47,7 @@ public class UserService {
             user.setName(user.getLogin());
         }
 
-        if (user.getLogin().chars().anyMatch(Character::isWhitespace)) {
+        if (!user.getLogin().chars().noneMatch(Character::isWhitespace)) {
             throw new InvalidFormatException("Логин не может содержать пробелов");
         }
 
@@ -56,12 +55,11 @@ public class UserService {
     }
 
     public User update(User newUser) {
-        if (newUser.getId() == null) {
-            throw new NotFoundException("Должен быть id");
+        if (newUser.getId() == null || userRepository.findById(newUser.getId()).isEmpty()) {
+            throw new NotFoundException("Пользователь с id = " + newUser.getId() + " не найден");
         }
-        throwIfNotExists(newUser.getId());
 
-        if (newUser.getLogin().chars().anyMatch(Character::isWhitespace)) {
+        if (!newUser.getLogin().chars().noneMatch(Character::isWhitespace)) {
             throw new InvalidFormatException("Логин не может содержать пробелов");
         }
 
@@ -69,39 +67,48 @@ public class UserService {
     }
 
     public void addFriend(Integer id, Integer friendId) {
-        if (id.equals(friendId)) {
-            log.warn("Попытка добавить себя в друзья: userId={}", id);
-            throw new IllegalArgumentException("Нельзя добавить себя в друзья");
+        if (userRepository.findById(id).isEmpty()) {
+            throw new NotFoundException("Пользователя с id = " + id + " нет");
         }
-        throwIfNotExists(id);
-        throwIfNotExists(friendId);
+        if (userRepository.findById(friendId).isEmpty()) {
+            throw new NotFoundException("Пользователя с id = " + friendId + " нет");
+        }
 
         userRepository.addFriend(id, friendId);
         log.info("User {} add user {} в друзья", id, friendId);
-        eventService.saveEvent(id, friendId, EventType.FRIEND, Operation.ADD);
+        Event event = new Event(); // добавление в ленту
+        event.setEventType(EventType.FRIEND.toString());
+        event.setUserId(id); // актор добавил в друзья
+        event.setEntityId(friendId);
+        event.setOperation(Operation.ADD.toString());
+        log.info("event внутри userService =" + event);
+        eventRepository.save(event); // добавление в ленту
     }
 
-    public List<User> removeFromFriends(Integer id, Integer friendId) {
-        throwIfNotExists(id);
-        throwIfNotExists(friendId);
-
-        eventService.saveEvent(id, friendId, EventType.FRIEND, Operation.REMOVE);
+    public Set<User> removeFromFriends(Integer id, Integer friendId) {
+        if (userRepository.findById(id).isEmpty()) {
+            throw new NotFoundException("Пользователя с id = " + id + " нет");
+        }
+        if (userRepository.findById(friendId).isEmpty()) {
+            throw new NotFoundException("Пользователя с id = " + friendId + " нет");
+        }
+        Event event = new Event(); // добавление в ленту
+        event.setEventType(EventType.FRIEND.toString());
+        event.setUserId(id); // актор удалил из друзей
+        event.setEntityId(friendId); // кого удалил
+        event.setOperation(Operation.REMOVE.toString());
+        eventRepository.save(event); // добавление в ленту
         return userRepository.removeFromFriends(id, friendId);
     }
 
-    public List<User> getFriendsToUser(Integer id) {
-        throwIfNotExists(id);
-
+    public Set<User> getFriendsToUser(Integer id) {
+        if (userRepository.findById(id).isEmpty()) {
+            throw new NotFoundException("Пользователя с id = " + id + " нет");
+        }
         return userRepository.getFriendsToUser(id);
     }
 
-    public List<User> getFriendsCommonOtherFriend(Integer id, Integer friendId) {
+    public Set<User> getFriendsCommonOtherFriend(Integer id, Integer friendId) {
         return userRepository.getFriendsCommonOtherFriend(id, friendId);
-    }
-
-    public void throwIfNotExists(Integer userId) {
-        if (!userRepository.existsById(userId)) {
-            throw new NotFoundException("Пользователь с id = " + userId + " не найден");
-        }
     }
 }
